requirements.txt
python-dotenv
streamlit
openai
google-cloud-texttospeech
requests
numpy
scipy
librosa
pystoi
pesq
speechrecognition
openai-whisper
jiwer
matplotlib
pydub
